Move express error handlers to dedicated file
